Refactoring registry instance generation
