+++ magnet-library/magnet-processor/src/main/java/magnet/processor/registry/RegistryProcessor.kt
@@ -28,7 +28,7 @@ class RegistryProcessor(
 ) {
 
     private val registryParser by lazy { RegistryParser(env) }
-    private val magnetIndexerGenerator by lazy { MagnetIndexerGenerator(env) }
+    private val magnetIndexerGenerator by lazy { RegistryGenerator(env) }
 
     private var generateRegistryOnNextRound = false
