+++ magnet-library/magnet-processor/src/main/java/magnet/processor/registry/instances/IndexGeneratorVisitor.kt
@@ -14,15 +14,15 @@
  * limitations under the License.
  */
 
-package magnet.processor.registry
+package magnet.processor.registry.instances
 
 import com.squareup.javapoet.ClassName
 import com.squareup.javapoet.CodeBlock
-import magnet.processor.registry.model.Index
-import magnet.processor.registry.model.IndexVisitor
-import magnet.processor.registry.model.Inst
-import magnet.processor.registry.model.Range
-import magnet.processor.registry.model.Section
+import magnet.processor.registry.instances.model.Index
+import magnet.processor.registry.instances.model.IndexVisitor
+import magnet.processor.registry.instances.model.Inst
+import magnet.processor.registry.instances.model.Range
+import magnet.processor.registry.instances.model.Section
 
 class IndexGeneratorVisitor : IndexVisitor {
