+++ magnet-library/magnet-processor/src/main/java/magnet/processor/registry/instances/model/IndexVisitor.kt
@@ -14,18 +14,10 @@
  * limitations under the License.
  */
 
-package magnet.processor.registry.model
+package magnet.processor.registry.instances.model
 
-class InstComparator : Comparator<Inst> {
-    override fun compare(left: Inst, right: Inst): Int {
-        val c1 = left.type.compareTo(right.type)
-        if (c1 != 0) {
-            return c1
-        }
-        val c2 = left.classifier.compareTo(right.classifier)
-        if (c2 != 0) {
-            return c2
-        }
-        return left.factory.compareTo(right.factory)
-    }
+interface IndexVisitor : InstVisitor {
+    fun visit(index: Index)
+    fun visit(section: Section)
+    fun visit(range: Range)
 }
