+++ magnet-library/magnet-processor/src/main/java/magnet/processor/registry/instances/model/InstVisitor.kt
@@ -14,26 +14,8 @@
  * limitations under the License.
  */
 
-package magnet.processor.registry.model
+package magnet.processor.registry.instances.model
 
-data class Range(
-    val type: String,
-    val classifier: String,
-    private val inst: Inst,
-    val from: Int
-) {
-    val impls = mutableListOf<Inst>()
-    val firstFactory
-        get() = impls[0].factory
-
-    init {
-        impls.add(inst)
-    }
-
-    fun accept(visitor: IndexVisitor) {
-        visitor.visit(this)
-        impls.forEach {
-            it.accept(visitor)
-        }
-    }
+interface InstVisitor {
+    fun visit(inst: Inst)
 }
