+++ magnet-library/magnet-processor/src/main/java/magnet/processor/registry/instances/model/Section.kt
@@ -14,18 +14,19 @@
  * limitations under the License.
  */
 
-package magnet.processor.registry.model
+package magnet.processor.registry.instances.model
 
-data class Index(
-    val instances: List<Inst>,
-    val sections: List<Section>
+data class Section(
+    val type: String
 ) {
+    val ranges = mutableMapOf<String, Range>()
+    val firstFactory
+        get() = ranges.values.elementAt(0).firstFactory
 
     fun accept(visitor: IndexVisitor) {
         visitor.visit(this)
-        sections.forEach {
-            it.accept(visitor)
+        ranges.forEach {
+            it.value.accept(visitor)
         }
     }
-
 }
